Refactor some fetcher stuff and fix some wearable bugs: - Changing PPU to 64 - Now able to properly get dimensions - Added a billboard script - Removing some hacks

+++ Assets/AavegotchiKit/Scripts/Gotchi/GotchiData.cs
@@ -7,8 +7,8 @@ namespace PortalDefender.AavegotchiKit
         public int hauntId;
         public string name;
         public string collateral;
-        public int[] numericTraits;
-        public int[] equippedWearables;
+        public short[] numericTraits;
+        public ushort[] equippedWearables;
         public int level;
         public int kinship;
         public int status; // 0 == portal, 1 == VRF_PENDING, 2 == open portal, 3 == Aavegotchi
@@ -20,8 +20,8 @@ namespace PortalDefender.AavegotchiKit
             hauntId = 1;
             name = "Default";
             collateral = "0xE0b22E0037B130A9F56bBb537684E6fA18192341";
-            numericTraits = new int[6];
-            equippedWearables = new int[8];
+            numericTraits = new short[6];
+            equippedWearables = new ushort[8];
             level = 1;
             kinship = 0;
             status = 0;
